refactor event logging to separate helper module
