+++ agent/lib/kontena/service_pods/common.rb
@@ -1,6 +1,9 @@
+require_relative '../helpers/event_log_helper'
+
 module Kontena
   module ServicePods
     module Common
+      include Kontena::Helpers::EventLogHelper
 
       # @param [String] service_id
       # @param [Integer] instance_number
@@ -18,20 +21,6 @@ module Kontena
         if container
           Docker::Container.get(container.id) rescue nil
         end
-      end
-
-      # @param [String] service_id
-      # @param [Integer] instance_number
-      # @param [String] reason
-      # @param [String] data
-      def log_service_pod_event(service_id, instance_number, reason, data, severity = Logger::INFO)
-        Celluloid::Notifications.publish('service_pod:event', {
-          service_id: service_id,
-          instance_number: instance_number,
-          severity: severity,
-          reason: reason,
-          data: data
-        })
       end
     end
   end
